Split linalg/det_and_inverse.hpp into linalg/det.hpp and linalg/inv.hpp
